+++ src/grid/grid/vertex.h
@@ -28,13 +28,13 @@ public:
 public:
     Vertices() {}
 
-    Vertices(int num_vertices) {
+    Vertices(size_t num_vertices) {
         _positions = vector_type("Vertices", num_vertices);
     }
 
     Vertices(vector_type positions) : _positions(positions) {}
 
-    void set_vertex_position(int vertex_id, Vector3<T> pos) {
+    void set_vertex_position(size_t vertex_id, Vector3<T> pos) {
         _positions(vertex_id, 0) = pos.x;
         _positions(vertex_id, 1) = pos.y;
         _positions(vertex_id, 2) = pos.z;
@@ -53,7 +53,7 @@ public:
     }
 
     KOKKOS_INLINE_FUNCTION
-    int size() const { return _positions.size(); }
+    size_t size() const { return _positions.size(); }
 
     mirror_type host_mirror() const {
         auto mirror_positions = _positions.host_mirror();
